+++ ogg_pager/CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Removed a bad assertion when writing nil packets ([PR](https://github.com/Serial-ATA/lofty-rs/pull/547))
+- Updated `MAX_CONTENT_SIZE` to the right value ([PR](https://github.com/Serial-ATA/lofty-rs/pull/600))
 
 ## [0.7.0] - 2025-01-05